src/requirements.txt
# Core dependencies
uagents>=0.8.0
psycopg2-binary>=2.9.6
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# For market data collection
yfinance>=0.2.28
python-binance>=1.0.17  # For crypto data
requests>=2.28.0

# For sentiment analysis
nltk>=3.8.1
textblob>=0.17.1
newsapi-python>=0.2.7  # For news data 